Add method for adding jobs
Since mysql2 appears to have some [problems](https://github.com/sidorares/node-mysql2/issues/939)
what comes to re-using the existing connection with new queries, the quick-and-dirty solution
was to open and close the connection for each query.

This is not ideal but works for now as the API testing is not very database-intensive job.